+++ src/assets/SelectableAssetBundle.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace davidhirtz\yii2\skeleton\assets;
+
+class SelectableAssetBundle extends AbstractAssetBundle
+{
+    public string $filename = 'selectable.js';
+}