+++ src/html/traits/AjaxAttributeTrait.php
@@ -31,7 +31,7 @@ trait AjaxAttributeTrait
         $new = clone $this;
 
         $new->attributes['hx-post'] = Url::to($url);
-        $new->attributes['hx-swap'] ??= 'show:top';
+        $new->attributes['hx-swap'] ??= 'outerHTML show:top';
         $new->attributes['hx-select'] ??= 'main';
         $new->attributes['hx-target'] ??= 'main';